requirements.txt
flask
reportlab
feedparser
google-generativeai
python-dotenv

